Arrumado persistência de novos usuarios

+++ src/main/java/com/montreal/portal/controler/UsuarioControler.java
@@ -27,7 +27,7 @@ public class UsuarioControler {
 	
 	@GetMapping(value = "/usuario/{codUsuario}")
 	@PreAuthorize("hasAnyRole('ADMIN','ANALISTA','TECNICO')")
-    public ResponseEntity<Response<Usuario>> findByCodUsuario(@PathVariable Long codUsuario) {
+    public ResponseEntity<Response<Usuario>> findByCodUsuario(@PathVariable Integer codUsuario) {
 		Response<Usuario> response = new Response<Usuario>();
 		Usuario users = usuarioRepository.findByCodUsuario(codUsuario);
 		if (users == null) {
